Generic validating framework added.
Basically the idea is that a set of validating functions will be mapped to specific API paths, and automatically executed.
If any of them return an error, we block the object provisioning.
The framework enables the easy expansion of the wbehook with other APIs... which we might need preeeeettyyyy sooon in the future ;)

Several functions in ipam and netcontrol packages are moved around to better fitting places.